add classifyImageObjects
