+++ convex/db/admin/see.ts
@@ -11,6 +11,10 @@ export const latestImages = query({
     const user = await ctx.viewerX()
     if (user.role !== 'admin') throw createError('Unauthorized', { code: 'unauthorized' })
 
-    return await ctx.table('images').order('desc').paginate(args.paginationOpts)
+    return await ctx
+      .table('images')
+      .order('desc')
+      .filter((q) => q.eq(q.field('deletionTime'), undefined))
+      .paginate(args.paginationOpts)
   },
 })
