+++ convex/workflows/actions/classifyImageObjects.ts
@@ -1,0 +1,41 @@
+import ky from 'ky'
+import * as vb from 'valibot'
+
+import { ENV } from '../../lib/env'
+import { getErrorMessage } from '../../shared/utils'
+
+const Response = vb.object({
+  modelId: vb.string(),
+  objects: vb.array(
+    vb.object({
+      label: vb.string(),
+      score: vb.number(),
+      box: vb.object({
+        xmin: vb.number(),
+        ymin: vb.number(),
+        xmax: vb.number(),
+        ymax: vb.number(),
+      }),
+    }),
+  ),
+})
+
+export const classifyImageObjects = async (args: { url: string }) => {
+  try {
+    const response = await ky
+      .get(ENV.WORKERS_IMAGE_CLASSIFICATION_URL, {
+        headers: {
+          Authorization: `Bearer ${ENV.WORKERS_API_KEY}`,
+        },
+        searchParams: {
+          image: args.url,
+        },
+      })
+      .json()
+
+    return vb.parse(Response, response)
+  } catch (err) {
+    console.error(err)
+    return { error: getErrorMessage(err) }
+  }
+}
